ahrc_webapp/backend/cervic_cancer/requirements.txt
tensorflow==2.17.1
fastapi==0.95.0
uvicorn==0.20.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.0
pydantic==1.10.13
typing-extensions==4.5.0
keras==3.2.0
h5py==3.10.0
